client: expose error::LightwayError
Error is also an interface of lib, such that we expose it as a mod.
Such that a developer using lightway-client crate will easy to debug and
follow by reading the source code.
